fix: updated BookList, Book and User classes to work inside the models directory. Passes the add test case.
