--- app/src/main/java/com/example/pocketbook/User.java
@@ -1,9 +1,0 @@
-package com.example.pocketbook;
-
-public class User {
-    private String userId;
-
-    public User(String userId) {
-        this.userId = userId;
-    }
-}
